+++ src/main/java/com/flansmod/common/driveables/CollisionBox.java
@@ -9,14 +9,14 @@ public class CollisionBox
 	/** The dimensions in model co-ordinates */
 	public float w, h, d;
 	/** The health of the part this box is attached to */
-	public int health;
+	public float health;
 	/** The part this box is attached to */
 	public EnumDriveablePart part;
 
 	/** The resistance of the part to penetration by bullets */
 	public float penetrationResistance = 5;
 	
-	public CollisionBox(int health, int x, int y, int z, int w, int h, int d)
+	public CollisionBox(float health, int x, int y, int z, int w, int h, int d)
 	{
 		this.health = health;
 		this.x = x / 16F;
@@ -27,7 +27,7 @@ public class CollisionBox
 		this.d = d / 16F;
 	}
 
-	public CollisionBox(int health, int x, int y, int z, int w, int h, int d, float penResistance)
+	public CollisionBox(float health, int x, int y, int z, int w, int h, int d, float penResistance)
 	{
 		this(health, x, y, z, w, h, d);
 		penetrationResistance = penResistance;
