+++ src/main/java/com/flansmod/common/driveables/ItemPlane.java
@@ -87,7 +87,7 @@ public class ItemPlane extends Item implements IPaintableItem
 		        NBTTagCompound tags = CompressedStreamTools.readCompressed(fileinputstream).getCompoundTag("data");
 		    	for(EnumDriveablePart part : EnumDriveablePart.values())
 		    	{
-		    		tags.setInteger(part.getShortName() + "_Health", type.health.get(part) == null ? 0 : type.health.get(part).health);
+		    		tags.setFloat(part.getShortName() + "_Health", type.health.get(part) == null ? 0 : type.health.get(part).health);
 		    		tags.setBoolean(part.getShortName() + "_Fire", false);
 		    	}
 		        fileinputstream.close();
@@ -239,7 +239,7 @@ public class ItemPlane extends Item implements IPaintableItem
     		tags.setString("Engine", PartType.defaultEngines.get(EnumType.plane).shortName);
     	for(EnumDriveablePart part : EnumDriveablePart.values())
     	{
-    		tags.setInteger(part.getShortName() + "_Health", type.health.get(part) == null ? 0 : type.health.get(part).health);
+    		tags.setFloat(part.getShortName() + "_Health", type.health.get(part) == null ? 0 : type.health.get(part).health);
     		tags.setBoolean(part.getShortName() + "_Fire", false);
     	}
     	planeStack.stackTagCompound = tags;
